Refactors config and argument handling
Improves config merging logic for migrations and seeds.

Renames `no-outdated-notice` argument to `no-outdated-check` for consistency.

Removes default value from `no-filesystem-caching` argument, allowing it to be explicitly controlled.

Improves transaction handling to avoid unnecessary transaction usage by default.
