+++ src/arguments/common.mts
@@ -3,7 +3,7 @@ import { CWDArg } from './cwd.mjs'
 import { DebugArg } from './debug.mjs'
 import { EnvironmentArg } from './environment.mjs'
 import { JitiArgs } from './jiti.mjs'
-import { NoOutdatedCheckArg } from './no-outdated-notice.mjs'
+import { NoOutdatedCheckArg } from './no-outdated-check.mjs'
 
 export const CommonArgs = {
 	...CWDArg,
