+++ src/arguments/jiti.mts
@@ -11,7 +11,6 @@ const ExperimentalResolveTSConfigPathsArg = {
 
 const NoFilesystemCachingArg = {
 	'no-filesystem-caching': {
-		default: false,
 		description:
 			'Will not write cache files to disk. See https://github.com/unjs/jiti#fscache for more information.',
 		type: 'boolean',
