+++ src/arguments/no-outdated-check.mts
@@ -2,7 +2,6 @@ import type { ArgsDef } from 'citty'
 
 export const NoOutdatedCheckArg = {
 	'no-outdated-check': {
-		default: false,
 		description:
 			'Will not check for latest kysely/kysely-ctl versions and notice newer versions exist.',
 		type: 'boolean',
